requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.3.0
moviepy==1.0.3